steam-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "steam-cli"
version = "0.1.0"
description = "Terminal-first Steam CLI for agents and humans"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
  "rich>=13.7,<14.0"
]

[project.scripts]
steam = "steam_cli.cli:run"

[tool.setuptools]
packages = ["steam_cli", "steam_cli.commands", "steam_cli.services", "steam_cli.providers", "steam_cli.models", "steam_cli.renderers", "steam_cli.launchers"]
